Изменения плейлистов, интерфейса, а также использование более новейших версий библиотек

+++ main.py
@@ -5,7 +5,7 @@ from mutagen.mp3 import MP3
 from mutagen import MutagenError
 from random import shuffle
 from PyQt5.QtWidgets import QApplication, QWidget, QMainWindow, QTextEdit, QFileDialog, \
-    QInputDialog, QAction, QLabel
+    QInputDialog, QLabel
 from PyQt5.QtCore import QUrl
 from PyQt5.QtGui import QPixmap
 from PyQt5.QtMultimedia import *
@@ -23,7 +23,7 @@ def time(ms):
 class Player(QMainWindow):
     def __init__(self):
         super().__init__()
-        uic.loadUi('audioplayer.ui', self) # Загрузка .ui файла
+        uic.loadUi('audioplayer.ui', self)  # Загрузка .ui файла
         # Инициализация
         self.player = QMediaPlayer()
         self.playlist = QMediaPlaylist()
@@ -42,7 +42,7 @@ class Player(QMainWindow):
         self.delete_track.triggered.connect(self.delete)
         self.check_text.triggered.connect(self.check_text_of_song)
         self.add_songs_to_playlist.clicked.connect(self.add_new_playlist)
-        self.delete_playlist.clicked.connect(self.del_playlist)
+        self.delete_playlist.clicked.connect(self.delete_playlist_from_list)
         self.mix_tracks.triggered.connect(self.mix)
         self.clear.triggered.connect(self.delete_all)
         self.play.clicked.connect(self.play_player)
@@ -54,6 +54,7 @@ class Player(QMainWindow):
         self.player.durationChanged.connect(self.update_duration)
         self.player.positionChanged.connect(self.update_position)
         self.time_slider.valueChanged[int].connect(self.change_pos)
+        self.playlists.currentIndexChanged.connect(self.download_playlist)
 
     def add(self, fnames):
         if not fnames:
@@ -68,23 +69,29 @@ class Player(QMainWindow):
                     self.playlist.addMedia(QMediaContent(QUrl.fromLocalFile(i)))
                 self.list_of_ways_to_files.extend(fnames)
         except MutagenError:
-            self.error = ErrorForm()
-            self.error.show()
+            error = ErrorForm()
+            error.show()
 
     def add_directory(self):
         # Диалоговое окно для выбора каталога
-        dirlist = QFileDialog.getExistingDirectory(self, "Выбрать папку", ".")
-        if dirlist != '':
-            all_files = os.listdir(dirlist) #Все файлы в выбранной директории
-            audiofiles = [dirlist + '/' + i for i in list(filter(lambda x: x.endswith('.mp3'), all_files))]
+        directory = QFileDialog.getExistingDirectory(self, "Выбрать папку", ".")
+        if directory != '':
+            # Получаем все файлы в выбранной директории
+            all_files = os.listdir(directory)
+            # Сортируем файлы на те, которые оканчиваются на .mp3
+            audiofiles = [directory + '/' + i for i in list(filter(lambda x: x.endswith('.mp3'), all_files))]
             self.add(audiofiles)
 
     def mix(self):
         # Перемешивание
+        # Очищаем MediaPlaylist, а также ListWidget
         self.playlist.clear()
         self.list_of_songs.clear()
+        # Перемешиваем список с путями к файлам
         shuffle(self.list_of_ways_to_files)
         for i in self.list_of_ways_to_files:
+            # Заново получаем все данные об аудиофайлах и наполняем
+            # MediaPlaylist и ListWidget
             self.list_of_songs.addItem(self.get_title(i))
             self.playlist.addMedia(QMediaContent(QUrl.fromLocalFile(i)))
         self.setWindowTitle('Audioplayer')
@@ -92,13 +99,16 @@ class Player(QMainWindow):
 
     def delete(self):
         # Удаление выбранных аудиофайлов из плейлиста
+        # Получаем список выбранных строк в ListWidget
         items = self.list_of_songs.selectedItems()
         for item in items:
+            # Удаляем выбранные строки из списка с путями к файлам,
+            # ListWidget и MediaPlaylist
             index = self.list_of_songs.row(item)
             self.list_of_songs.takeItem(index)
             self.playlist.removeMedia(index)
             del self.list_of_ways_to_files[index]
-            self.now_playing_track()
+        self.now_playing_track()
 
     def delete_all(self):
         # Удаление плейлиста полностью
@@ -132,10 +142,12 @@ class Player(QMainWindow):
 
     def change_vol(self):
         # Изменение громкости
+        # Громкость в player ставится в соответствии значению volume_slider'а
         self.player.setVolume(self.volume_slider.value())
 
     def update_duration(self, duration):
-        # Вывод длительности аудиофайла
+        # Установка максимального значения time_slider'а и вывод длительности
+        # аудиофайла на end_time (QLabel справа от time_slider)
         self.time_slider.setMaximum(duration)
         self.end_time.setText(time(duration))
 
@@ -152,6 +164,9 @@ class Player(QMainWindow):
     def now_playing_track(self):
         # Вывод всяческой информации о песне, которая играет в данный момент
         if self.playlist.isEmpty() or not self.player.isSeekable():
+            # Если сейчас ничего не играет или плейлист пустой, то
+            # длительность аудиофайла и заголовок окна сбрасываются, а
+            # обложка альбома прячется
             self.setWindowTitle('Audioplayer')
             self.end_time.setText('0:00')
             self.album_pic.hide()
@@ -163,9 +178,7 @@ class Player(QMainWindow):
 
     def get_title(self, file):
         # Получение  заголовка для окна
-        audio = MP3(file) #Считывание всех метаданных
-        '''TDRC (год), TALB (альбом), TIT2 (название трека),
-        TPE1 (исполнитель), TCON (жанр), COMM:XXX (текст), APIC (обложка альбома)'''
+        audio = MP3(file)  # Считывание всех метаданных
         if 'TIT2' in audio:
             title = str(audio['TIT2'])
         else:
@@ -178,7 +191,10 @@ class Player(QMainWindow):
 
     def check_info_about_song(self, file):
         # Получение всей основной информации о песне из метаданных
-        audio = MP3(file) # Считывание всех метаданных
+        audio = MP3(file)  # Считывание всех метаданных
+        # Ключи к метаданным: TDRC - год, TALB - альбом,
+        # TIT2 - название трека, TPE1 - исполнитель, TCON - жанр,
+        # COMM:XXX - текст, APIC - обложка альбома
         if 'TIT2' in audio:
             title = str(audio['TIT2'])
         else:
@@ -206,10 +222,10 @@ class Player(QMainWindow):
         # Получение обложки альбома из метаданных
         cover_name = 'cover.png'
         cover_key = ''
-        audio = MP3(track) # Считывание всех метаданных
+        audio = MP3(track)  # Считывание всех метаданных
         for i in audio.keys():
             if i.startswith('APIC'):
-                cover_key = i # Получение правильного ключа с обложкой
+                cover_key = i  # Получение правильного ключа с обложкой
         if cover_key != '':
             cover_binary = audio[cover_key]
             with open(cover_name, mode="wb") as cover:
@@ -233,56 +249,74 @@ class Player(QMainWindow):
 
     def add_new_playlist(self):
         # Добавление плейлиста
+        # Диалоговое окно с вводом названия плейлиста
         title, ok_pressed = QInputDialog.getText(
             self, "Плейлист", 'Введите название плейлиста')
         with open('playlists.txt', mode='a+', encoding='utf-8') as txt_of_playlists:
             if ok_pressed:
+                # Получение всех путей к файлам в виде строки, разделённой \n
                 all_files = "\n".join(self.list_of_ways_to_files)
+                # Запись в playlists.txt (в начале идёт название, потом все
+                # пути к аудиофайлам, разделённые \n, а в конце \n\n
                 txt_of_playlists.write(f'{title}\n{all_files}\n\n')
         self.check_playlists()
 
     def download_playlist(self):
         # Загрузка плейлиста на QListWidget
-        self.delete_all()
-        name = self.sender().text()
-        ind = self.list_of_names_of_playlists.index(name)
-        fnames = self.list_of_tracks_of_playlists[ind]
-        self.add(*fnames)
+        # self.playlists - ComboBox
+        name = self.playlists.currentText()  # Выбранный текст в ComboBox
+        if name != 'Плейлисты':  # "Плейлисты" - начальное значение ComboBox'а
+            # Удаление всех текущих аудиофайлов
+            self.delete_all()
+            # Получаем индекс данного плейлиста в списке с названиями плейлистов
+            ind = self.list_of_names_of_playlists.index(name)
+            # Получаем все пути к файлам и загружаем их
+            file_names = self.list_of_tracks_of_playlists[ind]
+            self.add(*file_names)
 
     def check_playlists(self):
         # Проверка playlists.txt на наличие плейлистов
         with open('playlists.txt', mode='r', encoding='utf-8') as txt_of_playlists:
+            # Считываем playlists.txt и получаем список из плейлистов (text)
             text = txt_of_playlists.read().split('\n\n')
             if text != ['']:
+                # Цикл всех плейлистов кроме последнего,
+                # так как в последнем пустая строка
                 for i in text[:-1]:
+                    # Разделяем все данные текущего плейлиста
                     playlist = i.split('\n')
+                    # Если данного названия плейлиста нету в списке текущих
+                    # загруженных плейлистов, то мы его загружаем
                     if playlist[0] not in self.list_of_names_of_playlists:
-                        self.tmp = QAction(self)
-                        self.tmp.setText(playlist[0])
-                        self.tmp.setObjectName(playlist[0])
-                        self.playlist_menu.addAction(self.tmp)
-                        self.tmp.triggered.connect(self.download_playlist)
+                        # Добавляем название плейлиста в ComboBox
+                        self.playlists.addItem(playlist[0])
+                        # Добавляем название плейлиста, а также его содержимое
+                        # в соответствующие плейлисты
                         self.list_of_names_of_playlists.append(playlist[0])
                         self.list_of_tracks_of_playlists.append([playlist[1:]])
 
-    def del_playlist(self):
+    def delete_playlist_from_list(self):
         # Удаление плейлиста
-        title, ok_pressed = QInputDialog.getText(
-            self, "Плейлист", 'Введите название плейлиста')
-        if ok_pressed:
+        name = self.playlists.currentText()  # Выбранный текст в ComboBox
+        if name != 'Плейлисты':  # "Плейлисты" - начальное значение ComboBox'а
+            # Получаем название плейлиста, который хотим удалить
+            title = self.list_of_names_of_playlists[
+                self.playlists.currentIndex() - 1]
             with open('playlists.txt', mode='a+', encoding='utf-8') as txt_of_playlists:
-                if title in self.list_of_names_of_playlists:
-                    text = list(filter(lambda x: not x.startswith(title),
-                                       txt_of_playlists.read().split('\n\n')))
-                    ind = self.list_of_names_of_playlists.index(title)
-                    del self.list_of_names_of_playlists[ind]
-                    del self.list_of_tracks_of_playlists[ind]
-                    self.playlist_menu.removeAction(self.findChild(QAction, title))
-                    txt_of_playlists.seek(0)
-                    txt_of_playlists.truncate()
-                    txt_of_playlists.write('\n\n'.join(text))
-                else:
-                    self.delete_playlist()
+                # Получаем плейлисты, исключая тот, который хотим удалить
+                text = list(filter(lambda x: not x.startswith(title),
+                                   txt_of_playlists.read().split('\n\n')))
+                # Получаем индекс плейлиста, который хотим удалить
+                ind = self.list_of_names_of_playlists.index(title)
+                # Удаляем плейлист из списков с данными о нём
+                del self.list_of_names_of_playlists[ind]
+                del self.list_of_tracks_of_playlists[ind]
+                # Удаляем плейлист из ComboBox
+                self.playlists.removeItem(ind + 1)
+                # Удаляем плейлист из playlists.txt
+                txt_of_playlists.seek(0)
+                txt_of_playlists.truncate()
+                txt_of_playlists.write('\n\n'.join(text))
 
 
 class TextForm(QWidget):
